+++ legal_knowledge_watch/models/__init__.py
@@ -4,4 +4,5 @@ from . import legal_watch
 from . import legal_watch_rule
 from . import legal_ingestion_run
 from . import legal_document_version
+from . import legal_dms_directory_route
 from . import legal_knowledge_document